Shuffle around command line handling functionality to reduce dependencies
- Remove dependency on command_line_handling from affinity_data
- Move command_line_handling helper functions to command_line_handling.{hpp,cpp}.
